@@ -29,8 +29,8 @@
     "jszip": "^3.10.1",
     "lodash-es": "^4.17.21",
     "multiformats": "^9.9.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-dropzone": "^14.2.3",
     "react-router-dom": "^6.4.2"
   },
