@@ -1,5 +1,4 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 
 import './index.css';
 
@@ -8,11 +7,13 @@ import { BrowserRouter } from 'react-router-dom';
 import { ShowPopupProvider } from './Components/Popups/Popups';
 import { DIDSign } from './Components/DidSign/DIDSign';
 
-ReactDOM.render(
+const root = document.getElementById('root');
+if (!root) throw new Error('Cannot find #root to render');
+
+createRoot(root).render(
   <BrowserRouter>
     <ShowPopupProvider>
       <DIDSign />
     </ShowPopupProvider>
   </BrowserRouter>,
-  document.getElementById('root'),
 );
